requirements.txt

langchain
langchain-core
langchain-openai
langchain-groq
openai
groq
pydantic
python-dotenv
pyyaml
rdflib
pandas